+++ src/GRA.Controllers/LookupController.cs
@@ -13,17 +13,21 @@ namespace GRA.Controllers
     public class LookupController : Base.Controller
     {
         private readonly ILogger<LookupController> _logger;
+        private readonly DynamicAvatarService _dynamicAvatarService;
         private readonly SchoolService _schoolService;
         private readonly SiteService _siteService;
         private readonly UserService _userService;
 
         public LookupController(ILogger<LookupController> logger,
              ServiceFacade.Controller context,
+             DynamicAvatarService dynamicAvatarService,
              SchoolService schoolService,
             SiteService siteService,
             UserService userService) : base(context)
         {
             _logger = Require.IsNotNull(logger, nameof(logger));
+            _dynamicAvatarService = Require.IsNotNull(dynamicAvatarService, 
+                nameof(dynamicAvatarService));
             _schoolService = Require.IsNotNull(schoolService, nameof(schoolService));
             _siteService = Require.IsNotNull(siteService, nameof(siteService));
             _userService = Require.IsNotNull(userService, nameof(userService));
@@ -83,6 +87,17 @@ namespace GRA.Controllers
         public async Task<JsonResult> UsernameInUse(string username)
         {
             return Json(await _userService.UsernameInUseAsync(username));
+        }
+
+        public async Task<JsonResult> GetItemsInBundleAsync(int id)
+        {
+            var bundle = await _dynamicAvatarService.GetBundleByIdAsync(id);
+            var thumbnailList = bundle.DynamicAvatarItems
+                .Select(_ => _pathResolver.ResolveContentPath(_.Thumbnail))
+                .ToList();
+
+
+            return Json(thumbnailList);
         }
     }
 }