+++ src/GRA.Data/Model/UserAvatarItem.cs
@@ -1,0 +1,15 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace GRA.Data.Model
+{
+    public class UserAvatarItem
+    {
+        [Required]
+        public int UserId { get; set; }
+        public User User { get; set; }
+
+        [Required]
+        public int DynamicAvatarItemId { get; set; }
+        public DynamicAvatarItem DynamicAvatarItem { get; set; }
+    }
+}