+++ src/GRA.Data/Model/UserLog.cs
@@ -20,5 +20,7 @@ namespace GRA.Data.Model
         public int? ChallengeId { get; set; }
         public int? BadgeId { get; set; }
         public string Description { get; set; }
+
+        public int? AvatarBundleId { get; set; }
     }
 }