+++ src/GRA.Domain.Repository/IDynamicAvatarItemRepository.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using GRA.Domain.Model;
 
@@ -6,6 +7,10 @@ namespace GRA.Domain.Repository
 {
     public interface IDynamicAvatarItemRepository : IRepository<DynamicAvatarItem>
     {
+        Task<ICollection<DynamicAvatarItem>> GetItemsByLayerAsync(int layerId);
         Task<ICollection<DynamicAvatarItem>> GetUserItemsByLayerAsync(int userId, int layerId);
+        Task<bool> HasUserUnlockedItemAsync(int userId, int itemId);
+        Task<ICollection<int>> GetUserUnlockedItemsAsync(int userId);
+        Task AddUserItemsAsync(int userId, List<int> itemId);
     }
 }